backend/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
pydantic-settings==2.10.1
python-dotenv==1.1.1
pytest==8.4.2
httpx==0.28.1
openai==1.107.0
twilio==9.8.3
websockets==15.0.1
debugpy==1.8.16
